auth_v1/requirements/test.txt
-r prod.txt
django-debug-toolbar==5.0.1
flake8==7.1.1
flake8-absolute-import==1.0.0.2
flake8-bugbear==24.12.12
flake8-builtins==2.5.0
flake8-commas==4.0.0
flake8-comments==0.1.2
flake8-dunder-all==0.4.1
flake8-expression-complexity==0.0.11
flake8-implicit-str-concat==0.5.0
flake8-import-order==0.18.2
flake8-pathlib==0.1.4
flake8-quotes==3.4.0
flake8-return==1.2.0
pep8-naming==0.14.1